+++ src/QuickMarcEditor/QuickMarcDuplicateWrapper.js
@@ -14,7 +14,10 @@ import getQuickMarcRecordStatus from './getQuickMarcRecordStatus';
 import {
   QUICK_MARC_ACTIONS,
 } from './constants';
-import { MARC_TYPES } from '../common/constants';
+import {
+  MARC_TYPES,
+  EXTERNAL_INSTANCE_APIS,
+} from '../common/constants';
 import {
   hydrateMarcRecord,
   removeFieldsForDuplicate,
@@ -52,6 +55,40 @@ const QuickMarcDuplicateWrapper = ({
   const showCallout = useShowCallout();
   const [httpError, setHttpError] = useState(null);
 
+  const saveLinksToNewRecord = async (externalId, marcRecord) => {
+    // request derived Instance record
+    const instancePromise = mutator.quickMarcEditInstance.GET({ path: `${EXTERNAL_INSTANCE_APIS[MARC_TYPES.BIB]}/${externalId}` });
+    // request derived MARC Bib record
+    const marcPromise = mutator.quickMarcEditMarcRecord.GET({ params: { externalId } });
+
+    Promise.all([instancePromise, marcPromise]).then(([{ _version }, derivedRecord]) => {
+      // copy linking data to new record
+      derivedRecord.fields = derivedRecord.fields.map((field) => {
+        // matching field from POST request
+        const matchingLinkedField = marcRecord.fields
+          .find(_field => _field.authorityId && _field.tag === field.tag && _field.authorityId === field.authorityId);
+
+        if (!matchingLinkedField) {
+          return field;
+        }
+
+        field.authorityNaturalId = matchingLinkedField.authorityNaturalId;
+        field.authorityControlledSubfields = matchingLinkedField.authorityControlledSubfields;
+
+        return field;
+      });
+
+      derivedRecord.relatedRecordVersion = parseInt(_version, 10);
+      mutator.quickMarcEditMarcRecord.PUT(derivedRecord)
+        .finally(() => {
+          history.push({
+            pathname: `/inventory/view/${externalId}`,
+            search: location.search,
+          });
+        });
+    });
+  };
+
   const onSubmit = useCallback(async (formValues) => {
     const formValuesToSave = removeDeletedRecords(formValues);
     const controlFieldErrorMessage = checkControlFieldLength(formValuesToSave);
@@ -86,19 +123,28 @@ const QuickMarcDuplicateWrapper = ({
     marcRecord.relatedRecordVersion = 1;
 
     return mutator.quickMarcEditMarcRecord.POST(marcRecord)
-      .then(({ qmRecordId }) => {
+      .then(async ({ qmRecordId }) => {
         history.push({
           pathname: '/inventory/view/id',
           search: location.search,
         });
 
-        getQuickMarcRecordStatus({
-          quickMarcRecordStatusGETRequest: mutator.quickMarcRecordStatus.GET,
-          qmRecordId,
-          showCallout,
-          history,
-          location,
-        });
+        try {
+          const { externalId } = await getQuickMarcRecordStatus({
+            quickMarcRecordStatusGETRequest: mutator.quickMarcRecordStatus.GET,
+            qmRecordId,
+            showCallout,
+          });
+
+          showCallout({ messageId: 'ui-quick-marc.record.saveNew.success' });
+
+          saveLinksToNewRecord(externalId, marcRecord);
+        } catch (e) {
+          showCallout({
+            messageId: 'ui-quick-marc.record.saveNew.error',
+            type: 'error',
+          });
+        }
       })
       .catch(async (errorResponse) => {
         const parsedError = await parseHttpError(errorResponse);