+++ src/queries/index.js
@@ -1,1 +1,2 @@
 export * from './useMarcSource';
+export * from './useAuthoritySourceFiles';